+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 <!-- here goes all the unreleased changes descriptions -->
 ### Fixes
 - fix crashing when of union's or interface type's `resolveType` function returns `undefined` or `null` (#731)
+- fix crashing when no reflected type available for fields with params decorators (#724)
 
 ## v1.1.0
 ### Features